% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_component_methods.R
\name{format.sf_choices}
\alias{format.sf_choices}
\title{Format an sf_choices object as a string}
\usage{
\method{format}{sf_choices}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sf_choices}.}

\item{...}{Ignored. Present for S3 consistency.}
}
\value{
A single character string.
}
\description{
Format an sf_choices object as a string
}
