% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_component_methods.R
\name{format.sf_scale}
\alias{format.sf_scale}
\title{Format an sf_scale object as a string}
\usage{
\method{format}{sf_scale}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sf_scale}.}

\item{...}{Ignored. Present for S3 consistency.}
}
\value{
A single character string.
}
\description{
Format an sf_scale object as a string
}
