% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_component_methods.R
\name{print.sf_branch}
\alias{print.sf_branch}
\title{Print an sf_branch object}
\usage{
\method{print}{sf_branch}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sf_branch}.}

\item{...}{Ignored. Present for S3 consistency.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an sf_branch object
}
\examples{
br <- sf_branch("q2", depends_on = "q1", operator = "==",
                value = "yes", action = "show")
print(br)
}
