% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_component_methods.R
\name{print.sf_choices}
\alias{print.sf_choices}
\title{Print an sf_choices object}
\usage{
\method{print}{sf_choices}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sf_choices}.}

\item{...}{Ignored. Present for S3 consistency.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an sf_choices object
}
\examples{
cs <- sf_choices("agree5", 1:5,
                 c("Strongly disagree", "Disagree", "Neutral",
                   "Agree", "Strongly agree"))
print(cs)
}
