% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_component_methods.R
\name{print.sf_model}
\alias{print.sf_model}
\title{Print an sf_model object}
\usage{
\method{print}{sf_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sf_model}.}

\item{...}{Ignored. Present for S3 consistency.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an sf_model object
}
