% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{sf_branches}
\alias{sf_branches}
\title{Get branching rules}
\usage{
sf_branches(x, ...)
}
\arguments{
\item{x}{A surveyframe object.}

\item{...}{Passed to methods.}
}
\value{
An \link{sf_component_list} of branching rules.
}
\description{
Returns the rules that control whether conditional items are shown.
}
\examples{
sf_branches(sframe_demo_data()$instrument)
}
\seealso{
\link{sf_accessors}
}
