% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{sf_checks}
\alias{sf_checks}
\title{Get response-quality checks}
\usage{
sf_checks(x, ...)
}
\arguments{
\item{x}{A surveyframe object.}

\item{...}{Passed to methods.}
}
\value{
An \link{sf_component_list} of declared checks.
}
\description{
Returns declared attention and other response-quality checks.
}
\examples{
sf_checks(sframe_demo_data()$instrument)
}
\seealso{
\link{sf_accessors}
}
