% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{sf_choice_sets}
\alias{sf_choice_sets}
\title{Get choice sets}
\usage{
sf_choice_sets(x, ...)
}
\arguments{
\item{x}{A surveyframe object.}

\item{...}{Passed to methods.}
}
\value{
An \link{sf_component_list} for an instrument or a choice-set data frame
for a codebook.
}
\description{
Returns the reusable value-and-label sets referenced by closed-response
items.
}
\examples{
sf_choice_sets(sframe_demo_data()$instrument)
}
\seealso{
\link{sf_accessors}
}
