% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_layer.R
\name{sf_covariance}
\alias{sf_covariance}
\title{Define a covariance between constructs}
\usage{
sf_covariance(from, to, label = NULL)
}
\arguments{
\item{from}{First construct ID.}

\item{to}{Second construct ID.}

\item{label}{Optional label.}
}
\value{
An object of class \code{sf_covariance}.
}
\description{
Define a covariance between constructs
}
\examples{
cov <- sf_covariance("sq", "sus", label = "cov1")
cov$from
cov$to
}
