% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{sf_id}
\alias{sf_id}
\title{Get an instrument component ID}
\usage{
sf_id(x, ...)
}
\arguments{
\item{x}{An \code{\link[=sf_item]{sf_item()}}, \code{\link[=sf_choices]{sf_choices()}}, \code{\link[=sf_scale]{sf_scale()}}, \code{\link[=sf_branch]{sf_branch()}},
\code{\link[=sf_check]{sf_check()}} or \code{\link[=sf_model]{sf_model()}} object.}

\item{...}{Passed to methods.}
}
\value{
A single character identifier.
}
\description{
Reads the stable identifier used to refer to a component elsewhere in the
instrument.
}
\examples{
sf_id(sf_item("q1", "Satisfaction", type = "numeric"))
}
\seealso{
\link{sf_identity}, \code{\link[=sf_label]{sf_label()}}
}
