% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{sf_items}
\alias{sf_items}
\title{Get survey items}
\usage{
sf_items(x, ...)
}
\arguments{
\item{x}{A surveyframe object.}

\item{...}{Passed to methods.}
}
\value{
An \link{sf_component_list} for an instrument or an item data frame for a
codebook.
}
\description{
Returns the declared question items in instrument order.
}
\examples{
sf_items(sframe_demo_data()$instrument)
}
\seealso{
\link{sf_accessors}
}
