% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{sf_meta}
\alias{sf_meta}
\title{Get survey metadata}
\usage{
sf_meta(x, ...)
}
\arguments{
\item{x}{A surveyframe object.}

\item{...}{Passed to methods.}
}
\value{
A list of instrument or codebook metadata.
}
\description{
Reads the title, version, description, language, validation state, and other
metadata without depending on the object's internal list layout.
}
\examples{
sf_meta(sframe_demo_data()$instrument)
}
\seealso{
\link{sf_accessors}
}
