% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{sf_models}
\alias{sf_models}
\title{Get model specifications}
\usage{
sf_models(x, ...)
}
\arguments{
\item{x}{A surveyframe object.}

\item{...}{Passed to methods.}
}
\value{
An \link{sf_component_list} for an instrument or a model data frame for a
codebook.
}
\description{
Returns declared CFA, SEM, PLS-SEM, mediation, and related model objects.
}
\examples{
sf_models(sframe_demo_data()$instrument)
}
\seealso{
\link{sf_accessors}
}
