% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{sf_plan}
\alias{sf_plan}
\title{Get the pre-declared analysis plan}
\usage{
sf_plan(x, ...)
}
\arguments{
\item{x}{A surveyframe object.}

\item{...}{Passed to methods.}
}
\value{
A list of analysis blocks for an instrument or a plan data frame for
a codebook.
}
\description{
Returns the ordered analysis blocks attached to an instrument or codebook.
}
\examples{
sf_plan(sframe_demo_data()$instrument)
}
\seealso{
\link{sf_accessors}, \link{sf_plan<-}
}
