% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{sf_scales}
\alias{sf_scales}
\title{Get survey scales}
\usage{
sf_scales(x, ...)
}
\arguments{
\item{x}{A surveyframe object.}

\item{...}{Passed to methods.}
}
\value{
An \link{sf_component_list} for an instrument or a scale data frame for a
codebook.
}
\description{
Returns the scale definitions, including their item membership and scoring
settings.
}
\examples{
sf_scales(sframe_demo_data()$instrument)
}
\seealso{
\link{sf_accessors}
}
