% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_component_methods.R
\name{summary.sf_branch}
\alias{summary.sf_branch}
\title{Summarise an sf_branch object}
\usage{
\method{summary}{sf_branch}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sf_branch}.}

\item{...}{Ignored. Present for S3 consistency.}
}
\value{
\code{object}, invisibly.
}
\description{
Summarise an sf_branch object
}
