% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_component_methods.R
\name{summary.sf_check}
\alias{summary.sf_check}
\title{Summarise an sf_check object}
\usage{
\method{summary}{sf_check}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sf_check}.}

\item{...}{Ignored. Present for S3 consistency.}
}
\value{
\code{object}, invisibly.
}
\description{
Summarise an sf_check object
}
