% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_component_methods.R
\name{summary.sf_choices}
\alias{summary.sf_choices}
\title{Summarise an sf_choices object}
\usage{
\method{summary}{sf_choices}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sf_choices}.}

\item{...}{Ignored. Present for S3 consistency.}
}
\value{
\code{object}, invisibly.
}
\description{
Summarise an sf_choices object
}
