% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_component_methods.R
\name{summary.sf_item}
\alias{summary.sf_item}
\title{Summarise an sf_item object}
\usage{
\method{summary}{sf_item}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sf_item}.}

\item{...}{Ignored. Present for S3 consistency.}
}
\value{
\code{object}, invisibly.
}
\description{
Summarise an sf_item object
}
