% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_component_methods.R
\name{summary.sf_model}
\alias{summary.sf_model}
\title{Summarise an sf_model object}
\usage{
\method{summary}{sf_model}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sf_model}.}

\item{...}{Ignored. Present for S3 consistency.}
}
\value{
\code{object}, invisibly.
}
\description{
Summarise an sf_model object
}
