% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_component_methods.R
\name{summary.sf_scale}
\alias{summary.sf_scale}
\title{Summarise an sf_scale object}
\usage{
\method{summary}{sf_scale}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sf_scale}.}

\item{...}{Ignored. Present for S3 consistency.}
}
\value{
\code{object}, invisibly.
}
\description{
Summarise an sf_scale object
}
