% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_layer.R
\name{add_model}
\alias{add_model}
\title{Add a model specification to an instrument}
\usage{
add_model(instrument, model, validate = TRUE, replace = TRUE)
}
\arguments{
\item{instrument}{An \code{sframe} object.}

\item{model}{An \code{\link[=sf_model]{sf_model()}} object.}

\item{validate}{Logical. Whether to validate the model against the
instrument before adding it.}

\item{replace}{Logical. Whether to replace an existing model with the same
ID. Defaults to \code{TRUE}.}
}
\value{
The updated \code{sframe} object.
}
\description{
Add a model specification to an instrument
}
