% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_layer.R
\name{efa_syntax}
\alias{efa_syntax}
\title{Generate EFA planning syntax}
\usage{
efa_syntax(
  items,
  nfactors = 1L,
  extraction = c("minres", "pa", "ml"),
  rotation = c("oblimin", "promax", "varimax"),
  data_name = "data"
)
}
\arguments{
\item{items}{Character vector of item IDs.}

\item{nfactors}{Number of factors.}

\item{extraction}{Extraction method.}

\item{rotation}{Rotation method.}

\item{data_name}{Name of the data object in generated R code.}
}
\value{
A character string with R syntax.
}
\description{
Generate EFA planning syntax
}
