% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_layer.R
\name{model_json}
\alias{model_json}
\title{Serialise a model specification to JSON}
\usage{
model_json(model, pretty = TRUE)
}
\arguments{
\item{model}{An \code{\link[=sf_model]{sf_model()}} object.}

\item{pretty}{Logical. Whether to pretty-print the JSON.}
}
\value{
A JSON string.
}
\description{
Serialise a model specification to JSON
}
