% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_layer.R
\name{seminr_syntax}
\alias{seminr_syntax}
\title{Generate seminr PLS-SEM syntax}
\usage{
seminr_syntax(model, data_name = "data", nboot = NULL, seed = 123)
}
\arguments{
\item{model}{An \code{\link[=sf_model]{sf_model()}} object of type \code{"pls_sem"}.}

\item{data_name}{Name of the data object in generated R code.}

\item{nboot}{Number of bootstrap samples.}

\item{seed}{Random seed for bootstrap syntax.}
}
\value{
An R syntax string for \code{seminr}.
}
\description{
Generate seminr PLS-SEM syntax
}
