% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{sf_accessors}
\alias{sf_accessors}
\alias{sf_meta}
\alias{sf_items}
\alias{sf_scales}
\alias{sf_choice_sets}
\alias{sf_branches}
\alias{sf_checks}
\alias{sf_models}
\alias{sf_plan}
\alias{sf_meta.sframe}
\alias{sf_items.sframe}
\alias{sf_scales.sframe}
\alias{sf_choice_sets.sframe}
\alias{sf_branches.sframe}
\alias{sf_checks.sframe}
\alias{sf_models.sframe}
\alias{sf_plan.sframe}
\alias{sf_meta.sframe_codebook}
\alias{sf_items.sframe_codebook}
\alias{sf_scales.sframe_codebook}
\alias{sf_choice_sets.sframe_codebook}
\alias{sf_models.sframe_codebook}
\alias{sf_plan.sframe_codebook}
\title{Explore a surveyframe object}
\usage{
sf_meta(x, ...)

sf_items(x, ...)

sf_scales(x, ...)

sf_choice_sets(x, ...)

sf_branches(x, ...)

sf_checks(x, ...)

sf_models(x, ...)

sf_plan(x, ...)

\method{sf_meta}{sframe}(x, ...)

\method{sf_items}{sframe}(x, ...)

\method{sf_scales}{sframe}(x, ...)

\method{sf_choice_sets}{sframe}(x, ...)

\method{sf_branches}{sframe}(x, ...)

\method{sf_checks}{sframe}(x, ...)

\method{sf_models}{sframe}(x, ...)

\method{sf_plan}{sframe}(x, ...)

\method{sf_meta}{sframe_codebook}(x, ...)

\method{sf_items}{sframe_codebook}(x, ...)

\method{sf_scales}{sframe_codebook}(x, ...)

\method{sf_choice_sets}{sframe_codebook}(x, ...)

\method{sf_models}{sframe_codebook}(x, ...)

\method{sf_plan}{sframe_codebook}(x, ...)
}
\arguments{
\item{x}{A surveyframe object.}

\item{...}{Passed to methods.}
}
\value{
\code{sf_items()}, \code{sf_scales()}, \code{sf_choice_sets()}, \code{sf_branches()},
\code{sf_checks()} and \code{sf_models()} return an \link{sf_component_list}.
\code{sf_meta()} and \code{sf_plan()} return lists.
}
\description{
Accessors for the parts of an instrument, a codebook, or a report. They
replace reaching into the object with \code{$}, which ties user code to the
internal layout.
}
\details{
\code{sf_items()}, \code{sf_scales()}, \code{sf_choice_sets()}, \code{sf_branches()},
\code{sf_checks()} and \code{sf_models()} return the component objects as an
\link{sf_component_list}. \code{sf_meta()} returns the metadata as a list and
\code{sf_plan()} returns the pre-declared analysis plan. For a flat table of the
same content, call \code{as.data.frame()} on the object instead.
}
\examples{
cs    <- sf_choices("ag5", 1:5,
           c("Strongly disagree", "Disagree", "Neutral",
             "Agree", "Strongly agree"))
item  <- sf_item("sat_1", "The service met my expectations.",
                 type = "likert", choice_set = "ag5", scale_id = "sat")
scale <- sf_scale("sat", "Satisfaction", items = "sat_1")
instr <- sf_instrument("Demo Survey", components = list(cs, item, scale))

sf_meta(instr)$title
sf_items(instr)
sf_scales(instr)[["sat"]]
as.data.frame(instr)
}
\seealso{
\code{\link[=as_sframe]{as_sframe()}}, \code{\link[=sf_problems]{sf_problems()}}, \link{sframe_validation}
}
