% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{sf_identity}
\alias{sf_identity}
\alias{sf_id}
\alias{sf_label}
\alias{sf_id.sf_item}
\alias{sf_id.sf_choices}
\alias{sf_id.sf_scale}
\alias{sf_id.sf_branch}
\alias{sf_id.sf_check}
\alias{sf_id.sf_model}
\alias{sf_label.sf_item}
\alias{sf_label.sf_choices}
\alias{sf_label.sf_scale}
\alias{sf_label.sf_branch}
\alias{sf_label.sf_check}
\alias{sf_label.sf_model}
\title{The ID and label of an instrument component}
\usage{
sf_id(x, ...)

sf_label(x, ...)

\method{sf_id}{sf_item}(x, ...)

\method{sf_id}{sf_choices}(x, ...)

\method{sf_id}{sf_scale}(x, ...)

\method{sf_id}{sf_branch}(x, ...)

\method{sf_id}{sf_check}(x, ...)

\method{sf_id}{sf_model}(x, ...)

\method{sf_label}{sf_item}(x, ...)

\method{sf_label}{sf_choices}(x, ...)

\method{sf_label}{sf_scale}(x, ...)

\method{sf_label}{sf_branch}(x, ...)

\method{sf_label}{sf_check}(x, ...)

\method{sf_label}{sf_model}(x, ...)
}
\arguments{
\item{x}{An \code{\link[=sf_item]{sf_item()}}, \code{\link[=sf_choices]{sf_choices()}}, \code{\link[=sf_scale]{sf_scale()}}, \code{\link[=sf_branch]{sf_branch()}},
\code{\link[=sf_check]{sf_check()}} or \code{\link[=sf_model]{sf_model()}} object.}

\item{...}{Passed to methods.}
}
\value{
A single character string. \code{sf_label()} returns \code{""} when the
component carries no label.
}
\description{
The ID and label of an instrument component
}
\examples{
item <- sf_item("q1", "How satisfied are you?", type = "likert",
                choice_set = "agree5")
sf_id(item)
sf_label(item)
}
