% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_builder.R
\name{sframe_builder_empty_state}
\alias{sframe_builder_empty_state}
\title{Create an empty SurveyStudio builder state}
\usage{
sframe_builder_empty_state()
}
\value{
A list containing empty metadata, choice, item, scale, branching,
and check collections suitable for SurveyStudio.
}
\description{
Create an empty SurveyStudio builder state
}
