% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sframe_plot_quality}
\alias{sframe_plot_quality}
\title{Quality report plot: straight-lining flag rate by scale}
\usage{
sframe_plot_quality(x, palette = c("web", "print"))
}
\arguments{
\item{x}{An \code{sframe_quality_report} object from \code{\link[=quality_report]{quality_report()}}.}

\item{palette}{One of \code{"web"} or \code{"print"}. See \code{sframe_brand()}.}
}
\value{
A ggplot2 object.
}
\description{
Quality report plot: straight-lining flag rate by scale
}
\seealso{
\code{\link[=quality_report]{quality_report()}}
}
