% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sframe_plot_reliability}
\alias{sframe_plot_reliability}
\title{Reliability plot: alpha and omega by scale}
\usage{
sframe_plot_reliability(x, palette = c("web", "print"))
}
\arguments{
\item{x}{An \code{sframe_reliability_report} object from \code{\link[=reliability_report]{reliability_report()}}.}

\item{palette}{One of \code{"web"} or \code{"print"}. See \code{\link[=sframe_brand]{sframe_brand()}}.}
}
\value{
A ggplot2 object.
}
\description{
Reliability plot: alpha and omega by scale
}
\seealso{
\code{\link[=reliability_report]{reliability_report()}}
}
