% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sframe_plot_validity}
\alias{sframe_plot_validity}
\title{Validity report plot: composite reliability and AVE by construct}
\usage{
sframe_plot_validity(x, palette = c("web", "print"))
}
\arguments{
\item{x}{An \code{sframe_validity_report} object from \code{\link[=validity_report]{validity_report()}}.}

\item{palette}{One of \code{"web"} or \code{"print"}. See \code{sframe_brand()}.}
}
\value{
A ggplot2 object.
}
\description{
Validity report plot: composite reliability and AVE by construct
}
\seealso{
\code{\link[=validity_report]{validity_report()}}
}
